pyproject.toml
[project]
name = "hedis-state-briefing-wall"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31",
  "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]
